+++ resources/views/common/reports/index.blade.php
@@ -3,9 +3,9 @@
 @section('title', trans_choice('general.reports', 2))
 
 @section('new_button')
-    @permission('create-common-reports')
+    @can('create-common-reports')
         <a href="{{ route('reports.create') }}" class="btn btn-success btn-sm">{{ trans('general.add_new') }}</a>
-    @endpermission
+    @endcan
     <a href="{{ route('reports.clear') }}" class="btn btn-warning btn-sm">{{ trans('general.clear_cache') }}</a>
 @endsection
 
@@ -26,14 +26,14 @@
                                 </a>
                                 <div class="dropdown-menu dropdown-menu-right dropdown-menu-arrow">
                                     <a class="dropdown-item" href="{{ route('reports.edit', $report->id) }}">{{ trans('general.edit') }}</a>
-                                    @permission('create-common-reports')
+                                    @can('create-common-reports')
                                         <div class="dropdown-divider"></div>
                                         <a class="dropdown-item" href="{{ route('reports.duplicate', $report->id) }}">{{ trans('general.duplicate') }}</a>
-                                    @endpermission
-                                    @permission('delete-common-reports')
+                                    @endcan
+                                    @can('delete-common-reports')
                                         <div class="dropdown-divider"></div>
                                         {!! Form::deleteLink($report, 'common/reports') !!}
-                                    @endpermission
+                                    @endcan
                                 </div>
                             </div>
                         </span>
