+++ resources/views/purchases/bills/index.blade.php
@@ -3,10 +3,10 @@
 @section('title', trans_choice('general.bills', 2))
 
 @section('new_button')
-    @permission('create-purchases-bills')
+    @can('create-purchases-bills')
         <a href="{{ route('bills.create') }}" class="btn btn-success btn-sm">{{ trans('general.add_new') }}</a>
         <a href="{{ route('import.create', ['group' => 'purchases', 'type' => 'bills']) }}" class="btn btn-white btn-sm">{{ trans('import.import') }}</a>
-    @endpermission
+    @endcan
     <a href="{{ route('bills.export', request()->input()) }}" class="btn btn-white btn-sm">{{ trans('general.export') }}</a>
 @endsection
 
@@ -73,21 +73,21 @@
                                             <div class="dropdown-divider"></div>
 
                                             @if ($item->status != 'cancelled')
-                                                @permission('create-purchases-bills')
+                                                @can('create-purchases-bills')
                                                     <a class="dropdown-item" href="{{ route('bills.duplicate', $item->id) }}">{{ trans('general.duplicate') }}</a>
-                                                @endpermission
+                                                @endcan
 
-                                                @permission('update-purchases-bills')
+                                                @can('update-purchases-bills')
                                                     <a class="dropdown-item" href="{{ route('bills.cancelled', $item->id) }}">{{ trans('general.cancel') }}</a>
-                                                @endpermission
+                                                @endcan
                                             @endif
 
-                                            @permission('delete-purchases-bills')
+                                            @can('delete-purchases-bills')
                                                 <div class="dropdown-divider"></div>
                                                 @if (!$item->reconciled)
                                                     {!! Form::deleteLink($item, 'bills.destroy') !!}
                                                 @endif
-                                            @endpermission
+                                            @endcan
                                         </div>
                                     </div>
                                 </td>
